proto/handshake: Move handshaking to p2p/src/proto/handshake.rs
Additionally, modify the handshake code to panic on unhandled
and invalid state/message combinations
